Move git.log.LogReverseEach to gitserver.Client
- move leftovers from git/log to gitdomain/log
