+++ internal/gitserver/gitdomain/log.go
@@ -1,14 +1,10 @@
-package git
+package gitdomain
 
 import (
 	"bufio"
-	"context"
 	"fmt"
 	"io"
 
-	"github.com/sourcegraph/sourcegraph/internal/api"
-	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -46,24 +42,6 @@ func LogReverseArgs(n int, givenCommit string) []string {
 		fmt.Sprintf("-%d", n),
 		givenCommit,
 	}
-}
-
-func LogReverseEach(repo string, db database.DB, commit string, n int, onLogEntry func(entry LogEntry) error) error {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	command := gitserver.NewClient(db).Command("git", LogReverseArgs(n, commit)...)
-	command.Repo = api.RepoName(repo)
-	// We run a single `git log` command and stream the output while the repo is being processed, which
-	// can take much longer than 1 minute (the default timeout).
-	command.DisableTimeout()
-	stdout, err := gitserver.StdoutReader(ctx, command)
-	if err != nil {
-		return err
-	}
-	defer stdout.Close()
-
-	return errors.Wrap(ParseLogReverseEach(stdout, onLogEntry), "ParseLogReverseEach")
 }
 
 func ParseLogReverseEach(stdout io.Reader, onLogEntry func(entry LogEntry) error) error {
